[RENAME] 커뮤니티 카테고리 전체 조회 파일 명 복수형으로 변경
