+++ functions/api/routes/community/communityCategoriesGET.js
